organized_database_creation/keyword_based_search/requirements.txt
# Enhanced Agriculture Data Curator Requirements
# Core dependencies
requests>=2.31.0
duckduckgo-search>=4.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# PDF processing dependencies
pypdf2>=3.0.0
pymupdf>=1.23.0

# OCR dependencies
pytesseract>=0.3.10
pillow>=10.0.0

# File type detection
python-magic>=0.4.27

# Configuration parsing
pyyaml>=6.0.1

# Additional utilities
urllib3>=2.0.0
charset-normalizer>=3.0.0

# New: Schema validation and JSON handling
jsonschema>=4.17.0

# New: Duplicate tracking and persistence
fcntl; sys_platform != 'win32'  # File locking (Unix-like systems only)

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
